Feature/send message

+++ internal/adapters/db/postgres/migrations/000003_chat.up.sql
@@ -2,14 +2,14 @@ CREATE TYPE "chat_t" AS enum ('private', 'group', 'channel');
 
 CREATE TABLE "chat"
 (
-  id        SERIAL PRIMARY KEY,
+  id        BIGSERIAL PRIMARY KEY,
   chat_type chat_t NOT NULL
 );
 
 CREATE TABLE "chat_users"
 (
-  chat_id INTEGER NOT NULL,
-  user_id INTEGER NOT NULL
+  chat_id BIGINT NOT NULL,
+  user_id BIGINT NOT NULL
 );
 
 ALTER TABLE
@@ -29,9 +29,9 @@ ALTER TABLE
 
 CREATE TABLE "messages"
 (
-  id         SERIAL PRIMARY KEY,
-  chat_id    INTEGER   NOT NULL,
-  sender_id  INTEGER   NOT NULL,
+  id         BIGSERIAL PRIMARY KEY,
+  chat_id    BIGINT   NOT NULL,
+  sender_id  BIGINT   NOT NULL,
   content    TEXT      NOT NULL,
   created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
