+++ internal/core/user.go
@@ -7,7 +7,7 @@ import (
 )
 
 type User struct {
-	ID        int32     `json:"id"`
+	ID        int64     `json:"id"`
 	Email     string    `json:"email"`
 	Phone     string    `json:"phone"`
 	FirstName string    `json:"first_name"`
@@ -18,7 +18,7 @@ type User struct {
 }
 
 type UserBySearch struct {
-	ID        int32  `json:"id"`
+	ID        int64  `json:"id"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
 }
@@ -32,6 +32,6 @@ type UserWithToken struct {
 
 type Session struct {
 	ID        uuid.UUID `json:"id"`
-	UserID    int32     `json:"user_id"`
+	UserID    int64     `json:"user_id"`
 	ExpiresAt time.Time `json:"expires_at"`
 }
